utils: move static delete fn to utils
