+++ src/v/utils/hdr_hist.h
@@ -18,7 +18,7 @@
 // vectorized types. needed comment to allow clang-format
 // header sorting to not resort cstdint
 #include "base/seastarx.h"
-#include "static_deleter_fn.h"
+#include "utils/static_deleter_fn.h"
 
 #include <seastar/core/metrics_types.hh>
 #include <seastar/core/temporary_buffer.hh>
