+++ src/v/utils/static_deleter_fn.h
